feat: handled select in a form in it's recommended way
